Доделан фронт

+++ guitar-shop-backend-nx-mono/apps/bff/src/app/products/products.controller.ts
@@ -1,32 +1,25 @@
 import axios from 'axios';
 import { Request } from 'express';
 
-import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Logger, LoggerService, Param, ParseUUIDPipe, Post, Put, Req, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, HttpCode, HttpStatus, Logger, LoggerService, Param, ParseUUIDPipe, Post, Put, Req, UploadedFile, UseGuards, UseInterceptors } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 
 import { FindGuitarProductCardsQuery, GuitarShopCreateProductCardBffDto, GuitarShopCreateProductCardDto, GuitarShopProductCardRdo, GuitarShopUpdateProductCardBffDto, JwtPayloadDto, TransformAndValidateDtoInterceptor, TransformAndValidateQueryInterceptor } from '@guitar-shop/shared-types';
 import { BffEnvInterface } from '../../assets/interface/bff-env.interface';
 import { fillDTO } from '@guitar-shop/core';
 import { FileInterceptor } from '@nestjs/platform-express';
-import { resolve } from 'path';
-import { MulterOptions } from '@nestjs/platform-express/multer/interfaces/multer-options.interface';
 import { JwtAuthUsersGuard } from '../../assets/guard/jwt-auth-users.guard';
 
 @Controller('products')
 export class ProductsController {
-  that = this;
   private readonly logger: LoggerService = new Logger(ProductsController.name);
 
   private readonly productsMicroserviceUrl: string;
-  private multerOptions: MulterOptions;
 
   constructor(
     private readonly config: ConfigService<BffEnvInterface>,
   ) {
     this.productsMicroserviceUrl = `http://${config.get("PRODUCTS_MICROSERVICE_HOST")}:${config.get("PRODUCTS_MICROSERVICE_PORT")}`;
-    this.multerOptions = {
-      dest: resolve('apps', 'bff', `${this.config.get('UPLOAD_DIR')}`),
-    };
   }
 
 
@@ -40,7 +33,6 @@ export class ProductsController {
   }
 
   @Get('/:productId')
-  @UseInterceptors(new TransformAndValidateQueryInterceptor(FindGuitarProductCardsQuery))
   @HttpCode(HttpStatus.OK)
   async findProductCardById(@Req() req: Request, @Param('productId', ParseUUIDPipe) _productId: string): Promise<GuitarShopProductCardRdo> {
     const result = (await axios.get(`${this.productsMicroserviceUrl}${req.url}`)).data as GuitarShopProductCardRdo;
@@ -48,14 +40,17 @@ export class ProductsController {
     return fillDTO(GuitarShopProductCardRdo, result);
   }
 
-
   @Post('/')
   @UseInterceptors(new TransformAndValidateDtoInterceptor(GuitarShopCreateProductCardBffDto))
   @UseInterceptors(FileInterceptor('image'))
   @UseGuards(JwtAuthUsersGuard)
   @HttpCode(HttpStatus.CREATED)
   async createProductCard(@Req() req: Request & { user: JwtPayloadDto, }, @Body() dto: GuitarShopCreateProductCardBffDto, @UploadedFile() file: Express.Multer.File): Promise<GuitarShopProductCardRdo> {
-    const { id } = req.user;
+    const { id, isAdmin } = req.user;
+
+    if (!isAdmin) {
+      throw new ForbiddenException('Доступ запрещен! Вы не являетесь администратором!');
+    }
 
     const dtoToProductsMicroservice: GuitarShopCreateProductCardDto = {
       creatorUserId: id,
@@ -76,6 +71,12 @@ export class ProductsController {
   @UseGuards(JwtAuthUsersGuard)
   @HttpCode(HttpStatus.OK)
   async updateProductCard(@Req() req: Request & { user: JwtPayloadDto, }, @Param('productId', ParseUUIDPipe) _productId: string, @Body() dto: GuitarShopUpdateProductCardBffDto, @UploadedFile() file: Express.Multer.File): Promise<GuitarShopProductCardRdo> {
+    const { isAdmin } = req.user;
+
+    if (!isAdmin) {
+      throw new ForbiddenException('Доступ запрещен! Вы не являетесь администратором!');
+    }
+
     if (file) {
       dto.imageLink = `/api/products/${this.config.get('UPLOAD_DIR')}/${file.filename}`;
     }
@@ -89,6 +90,12 @@ export class ProductsController {
   @UseGuards(JwtAuthUsersGuard)
   @HttpCode(HttpStatus.OK)
   async deleteProductCardById(@Req() req: Request & { user: JwtPayloadDto, }, @Param('productId', ParseUUIDPipe) _productId: string): Promise<string> {
+    const { isAdmin } = req.user;
+
+    if (!isAdmin) {
+      throw new ForbiddenException('Доступ запрещен! Вы не являетесь администратором!');
+    }
+
     return (await axios.delete(`${this.productsMicroserviceUrl}${req.url}`)).data as string;
   }
